Create README.md
# nodeBasicAuthCodeforAPIs
This is the code I used, a very basic approach to have a successful response from an API that requires Basic-Authentication

This is good for node.js and requires no packages and is for basic-authentication when it is required for a get rest call.
I have also entered a file that may help you when you are trying to iterate through JSON data.  I have added notes to both to try
and help guide you through it as you are attempting to use it.  I am kind of new to this myself, so am hoping this helps someone
else.